requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
alabaster==0.7.12
    # via sphinx
asgiref==3.5.0
    # via django
babel==2.9.1
    # via sphinx
beautifulsoup4==4.10.0
    # via -r requirements.in
bleach==4.1.0
    # via -r requirements.in
boto3==1.21.36
    # via -r requirements.in
botocore==1.24.36
    # via
    #   boto3
    #   s3transfer
brotli==1.0.9
    # via
    #   fonttools
    #   whitenoise
certifi==2019.11.28
    # via requests
cffi==1.14.3
    # via
    #   cryptography
    #   weasyprint
charset-normalizer==2.0.4
    # via requests
cryptography==3.3.2
    # via pyjwt
cssselect2==0.4.1
    # via weasyprint
defusedxml==0.6.0
    # via python3-openid
dj-database-url==0.5.0
    # via -r requirements.in
dj-stripe==2.6.1
    # via -r requirements.in
django==4.0.3
    # via
    #   -r requirements.in
    #   dj-stripe
    #   django-allauth
    #   django-anymail
    #   django-debug-toolbar
    #   django-extensions
    #   django-hashid-field
    #   django-hijack
    #   django-htmx
    #   django-storages
    #   django-tz-detect
    #   jsonfield
django-allauth==0.50.0
    # via -r requirements.in
django-anymail[sendgrid]==8.5
    # via -r requirements.in
django-debug-toolbar==3.2.4
    # via -r requirements.in
django-environ==0.8.1
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
django-hashid-field==3.3.4
    # via -r requirements.in
django-hijack==3.2.0
    # via -r requirements.in
django-htmx==1.9.0
    # via -r requirements.in
django-ordered-model==3.5
    # via -r requirements.in
django-simple-history==3.0.0
    # via -r requirements.in
django-storages==1.12.3
    # via -r requirements.in
django-tz-detect==0.4.0
    # via -r requirements.in
django-waffle==2.4.1
    # via -r requirements.in
docutils==0.16
    # via
    #   sphinx
    #   sphinx-rtd-theme
fonttools[woff]==4.28.5
    # via weasyprint
gunicorn==20.1.0
    # via -r requirements.in
hashids==1.3.1
    # via django-hashid-field
html5lib==1.1
    # via weasyprint
idna==2.8
    # via requests
imagesize==1.2.0
    # via sphinx
importlib-metadata==4.11.3
    # via sphinx
jinja2==2.11.3
    # via sphinx
jmespath==1.0.0
    # via
    #   boto3
    #   botocore
jsonfield==3.1.0
    # via dj-stripe
markupsafe==1.1.1
    # via jinja2
oauthlib==3.1.0
    # via requests-oauthlib
packaging==20.4
    # via
    #   bleach
    #   sphinx
pillow==9.0.1
    # via weasyprint
psycopg2==2.9.3
    # via -r requirements.in
pycparser==2.20
    # via cffi
pydyf==0.1.2
    # via weasyprint
pygments==2.7.4
    # via sphinx
pyjwt[crypto]==1.7.1
    # via django-allauth
pyparsing==2.4.7
    # via packaging
pyphen==0.12.0
    # via weasyprint
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   botocore
    #   time-machine
python3-openid==3.1.0
    # via django-allauth
pytz==2019.3
    # via
    #   babel
    #   django-tz-detect
requests==2.27.1
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
    #   rollbar
    #   sphinx
    #   stripe
requests-oauthlib==1.3.0
    # via django-allauth
rollbar==0.16.2
    # via -r requirements.in
s3transfer==0.5.2
    # via boto3
six==1.14.0
    # via
    #   bleach
    #   cryptography
    #   html5lib
    #   packaging
    #   python-dateutil
    #   rollbar
    #   sphinx-sitemap
snowballstemmer==2.0.0
    # via sphinx
soupsieve==1.9.5
    # via beautifulsoup4
sphinx==4.5.0
    # via
    #   -r requirements.in
    #   sphinx-rtd-theme
    #   sphinx-sitemap
sphinx-rtd-theme==1.0.0
    # via -r requirements.in
sphinx-sitemap==2.2.0
    # via -r requirements.in
sphinxcontrib-applehelp==1.0.2
    # via sphinx
sphinxcontrib-devhelp==1.0.2
    # via sphinx
sphinxcontrib-htmlhelp==2.0.0
    # via sphinx
sphinxcontrib-jsmath==1.0.1
    # via sphinx
sphinxcontrib-qthelp==1.0.3
    # via sphinx
sphinxcontrib-serializinghtml==1.1.5
    # via sphinx
sqlparse==0.3.0
    # via
    #   django
    #   django-debug-toolbar
stripe==2.55.2
    # via dj-stripe
time-machine==2.6.0
    # via -r requirements.in
tinycss2==1.1.1
    # via
    #   cssselect2
    #   weasyprint
types-bleach==4.1.5
    # via -r requirements.in
types-python-dateutil==2.8.10
    # via -r requirements.in
types-pytz==2021.3.6
    # via -r requirements.in
types-requests==2.27.16
    # via -r requirements.in
types-six==1.16.12
    # via -r requirements.in
types-toml==0.10.4
    # via -r requirements.in
types-urllib3==1.26.7
    # via types-requests
urllib3==1.26.5
    # via
    #   botocore
    #   requests
weasyprint==54.2
    # via -r requirements.in
webencodings==0.5.1
    # via
    #   bleach
    #   cssselect2
    #   html5lib
    #   tinycss2
whitenoise[brotli]==6.0.0
    # via -r requirements.in
zipp==3.8.0
    # via importlib-metadata
zopfli==0.1.9
    # via fonttools

# The following packages are considered to be unsafe in a requirements file:
# setuptools
